backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.5
google-adk>=1.3.0  # Latest ADK with improved agent and streaming support

# Cloud services (optional)
google-cloud-aiplatform>=1.36.0
requests>=2.30.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0